Several fixes (#8)
* remove debug/test code which sneaked in

* use urllib to properly encode request URL

The urllib makes sure the get parameters are properly encoded.

* get the actual next three connection (fixes #7)

Slicing with [1:4] returns connection at index 1, 2, 3 but misses the
first connection at index 0.

* move connection dictionary creation to separate function

* skip sections where journey is null

Some connections have journy set to null, e.g. when there is a walking
part to make the connecton, e.g.
  curl http://transport.opendata.ch/v1/connections\?from\=Z%C3%BCrich%2C+Blumenfeldstrasse\&to\=Z%C3%BCrich+Oerlikon%2C+Bahnhof \
    | jq ".connections[1].sections[0].journey"

Skip those sections

* add parameter to limit number of connections returned

The API supports the limit parameter to limit the amount of connections
returned. Make use of it and set limit by default to 3 (previous limit).
Also add a parameter so the use can select the amount of conenctions
returned.